Code review: docs and int -> long

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/newapi/PartitionedNodeCursorScan.java
@@ -45,7 +45,7 @@ final class PartitionedNodeCursorScan implements PartitionedScan<NodeCursor> {
         } else {
             this.numberOfPartitions = (int) totalCount;
         }
-        this.batchSize = (int) Math.ceil((double) totalCount / numberOfPartitions);
+        this.batchSize = (long) Math.ceil((double) totalCount / numberOfPartitions);
         this.emittedPartitions = new AtomicInteger(0);
     }
